+++ Excel_PRIME.Tests/PerformanceTesting.cs
@@ -13,10 +13,10 @@ namespace ExcelPRIME.Tests;
 internal class PerformanceTesting
 {
     [Test]
-    [TestCase("Data/sampledocs-50mb-xlsx-file.xlsx", 7000014)]
-    [TestCase("Data/Blank Data 1 Million Rows.xlsx", 25601276)]
-    [TestCase("Data/sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
-    [TestCase("Data/100mb.xlsx", 8935680)]
+    [TestCase("sampledocs-50mb-xlsx-file.xlsx", 7000012)]
+    [TestCase("Blank Data 1 Million Rows.xlsx", 14314945)]
+    [TestCase("sampledocs-50mb-xlsx-file-sst.xlsx", 7000012)]
+    [TestCase("100mb.xlsx", 8930256)]
     [Explicit("Lot of data being thrown about !")]
     public async Task A010_AccessEveryCellExcel_Prime(string fileName, int expectedCells)
     {
@@ -26,10 +26,10 @@ internal class PerformanceTesting
     }
 
     [Test]
-    [TestCase("Data/sampledocs-50mb-xlsx-file.xlsx", 7000014)]
-    [TestCase("Data/Blank Data 1 Million Rows.xlsx", 25601276)]
-    [TestCase("Data/sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
-    [TestCase("Data/100mb.xlsx", 8935680)]
+    [TestCase("sampledocs-50mb-xlsx-file.xlsx", 7000014)]
+    [TestCase("Blank Data 1 Million Rows.xlsx", 25601276)]
+    [TestCase("sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
+    [TestCase("100mb.xlsx", 8935680)]
     [Explicit("Lot of data being thrown about !")]
     public void A020_AccessEveryCellXlsxHelper(string fileName, int expectedCells)
     {
@@ -39,10 +39,10 @@ internal class PerformanceTesting
     }
 
     [Test]
-    [TestCase("Data/sampledocs-50mb-xlsx-file.xlsx", 7000014)]
-    [TestCase("Data/Blank Data 1 Million Rows.xlsx", 25601276)]
-    [TestCase("Data/sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
-    [TestCase("Data/100mb.xlsx", 8935680)]
+    [TestCase("sampledocs-50mb-xlsx-file.xlsx", 7000014)]
+    [TestCase("Blank Data 1 Million Rows.xlsx", 25601276)]
+    [TestCase("sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
+    [TestCase("100mb.xlsx", 8935680)]
     [Explicit("Lot of data being thrown about !")]
     public async Task A030_AccessEveryCellSylvan(string fileName, int expectedCells)
     {
@@ -52,10 +52,10 @@ internal class PerformanceTesting
     }
 
     //[Test]
-    //[TestCase("Data/sampledocs-50mb-xlsx-file.xlsx", 7000014)]
-    //[TestCase("Data/Blank Data 1 Million Rows.xlsx", 25601276)]
-    //[TestCase("Data/sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
-    //[TestCase("Data/100mb.xlsx", 8935680)]
+    //[TestCase("sampledocs-50mb-xlsx-file.xlsx", 7000014)]
+    //[TestCase("Blank Data 1 Million Rows.xlsx", 25601276)]
+    //[TestCase("sampledocs-50mb-xlsx-file-sst.xlsx", 7000014)]
+    //[TestCase("100mb.xlsx", 8935680)]
     //[Explicit("Lot of data being thrown about !")]
     //public void A040_AccessEveryCellFastExcel(string fileName, int expectedCells)
     //{
@@ -63,4 +63,17 @@ internal class PerformanceTesting
     //    int cells = aecB.AccessEveryCellFastExcel();
     //    cells.Should().Be(expectedCells);
     //}
+
+    [Test]
+    [TestCase("sampledocs-50mb-xlsx-file.xlsx", 7000012)]
+    [TestCase("Blank Data 1 Million Rows.xlsx", 14314945)]
+    [TestCase("sampledocs-50mb-xlsx-file-sst.xlsx", 7000012)]
+    [TestCase("100mb.xlsx", 8930256)]
+    [Explicit("Lot of data being thrown about !")]
+    public async Task A040_ParallelEveryCellAsyncExcel_PrimeTwice(string fileName, int expectedCells)
+    {
+        AccessEveryCellBenchmarks aecB = new AccessEveryCellBenchmarks { FileName = fileName };
+        int cells = await aecB.ParallelEveryCellAsyncExcel_PrimeTwice();
+        cells.Should().Be(expectedCells);
+    }
 }